Agriculture Assistant agent/requirements.txt
requests
python-dotenv
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
joblib>=1.3.0
groq>=0.4.0
opencv-python>=4.8.0
